split inspector
